requirements.txt
pyyaml
rx
pyxb
toolz